+++ Modules/Sources/Core/Engine/ArchiveSwcEngine.swift
@@ -1,0 +1,67 @@
+//
+//  ArchiveSwcEngine.swift
+//  Modules
+//
+//  Created by Stephan Arenswald on 14.12.25.
+//
+
+import Foundation
+import SWCompression
+
+final class ArchiveSwcEngine: ArchiveEngine {
+    private func stripFileExtension( _ filename: String ) -> String {
+        var components = filename.components(separatedBy: ".")
+        guard components.count > 1 else { return filename }
+        components.removeLast()
+        return components.joined(separator: ".")
+    }
+    
+    func loadArchive(url: URL) async throws -> [ArchiveItem] {
+        let name = stripFileExtension(url.lastPathComponent)
+        return [
+            ArchiveItem(name: String(name), virtualPath: name, type: .file)
+        ]
+    }
+    
+    func extract(item: ArchiveItem, from url: URL, to destination: URL) async throws -> URL? {
+        let sourceFileName = url.lastPathComponent
+        let extractedFileName = stripFileExtension(sourceFileName)
+        let extractedFilePathName = url.appendingPathComponent(extractedFileName, isDirectory: false)
+        
+        do {
+            if let data = try? Data(contentsOf: url, options: .mappedIfSafe) {
+                let decompressedData = try LZ4.decompress(data: data)
+                
+                FileManager.default.createFile(atPath: extractedFilePathName.path, contents: decompressedData)
+                
+                return extractedFilePathName
+            } else {
+                Logger.error("Could not decompress archive")
+            }
+        } catch {
+            Logger.error(error.localizedDescription)
+        }
+        
+        return nil
+    }
+    
+    func extract(_ url: URL, to destination: URL) async throws {
+        let sourceFileName = url.lastPathComponent
+        let extractedFileName = stripFileExtension(sourceFileName)
+        let extractedFilePathName = url.appendingPathComponent(extractedFileName, isDirectory: false)
+        
+        do {
+            if let data = try? Data(contentsOf: url, options: .mappedIfSafe) {
+                let decompressedData = try LZ4.decompress(data: data)
+                
+                FileManager.default.createFile(atPath: extractedFilePathName.path, contents: decompressedData)
+            } else {
+                Logger.error("Could not decompress archive")
+            }
+        } catch {
+            Logger.error(error.localizedDescription)
+        }
+    }
+    
+    
+}